Aim 2/requirements.txt
ebmlite==1.0.1
numpy==1.16.2
pyserial==3.4
